almost done refactoring
