+++ app/routes/auth.py
@@ -1,0 +1,45 @@
+from flask import render_template, redirect, url_for, flash, Response
+from flask_login import current_user, login_user, logout_user
+from app.models import User
+from app.forms import LoginForm, RegistrationForm
+from . import endpoint
+
+
+@endpoint.route("/login", methods=["GET", "POST"])
+def login() -> str | Response:
+    if current_user.is_authenticated:
+        return redirect(url_for("routes.index"))
+    form = LoginForm()
+    if form.validate_on_submit():
+        user = User.query.filter_by(username=form.username.data).first()
+        if user is None or not user.check_password(form.password.data):
+            flash("Invalid username or password")
+            return redirect(url_for("routes.login"))
+        login_user(user, remember=form.remember_me.data)
+        return redirect(url_for("routes.index"))
+    return render_template("login.html", title="Sign In", form=form)
+
+
+@endpoint.route("/logout")
+def logout() -> Response:
+    logout_user()
+    return redirect(url_for("routes.index"))
+
+
+@endpoint.route("/register", methods=["GET", "POST"])
+def register() -> str | Response:
+    if current_user.is_authenticated:
+        return redirect(url_for("routes.index"))
+    form = RegistrationForm()
+    if form.validate_on_submit():
+        user = User(username=form.username.data, email=form.email.data)
+        user.set_password(form.password.data)
+        first_admin = User.query.filter_by(is_admin=True).first()
+        if not first_admin:
+            user.is_admin = True
+        user.save()
+        flash(f"New user {form.username.data} has been created!")
+        if User.query.count() == 1:
+            flash("You are the first user, so you are an admin.")
+        return redirect(url_for("routes.login"))
+    return render_template("register.html", title="Register", form=form)
